refactor : 투표, 댓글 selector 변경
