+++ src/hooks/useSelectedComment.ts
@@ -27,23 +27,3 @@ export const useSelectedComment = () => {
 
   return filteredComments;
 };
-
-export const useSelectedVote = () => {
-  const comments = useSelector(
-    (state: RootState) => state.postDetail.post.comments,
-  );
-
-  const filteredVotes = useMemo(() => {
-    if (!comments) {
-      return [];
-    } else {
-      return comments.filter((comment) => {
-        const parsedComment = JSON.parse(comment.comment) as Comment;
-
-        return parsedComment.type === 'vote';
-      });
-    }
-  }, [comments]);
-
-  return filteredVotes;
-};
